requirements-bhmk.txt
pyserial>=3.5
protobuf>=3.20.0,<3.21
markdown>=3.10.2
beautifulsoup4>=4.14.3
PyYAML>=6.0.3
rich>=14.3.3
pandas
typer>=0.24.1
structlog>=25.5.0
tenacity>=9.1.4
